Alpha 0.2
CONQUEST.JAVA CHANGES
Reorganised code to make it more readable and easier to modify.
Changed window size to 500x300 (from 600, 400).
Made initControls() method a higher level method in order to oversee the
many steps require to initialise and set control buttons in their
intended locations. As a result of this, ActionListener creation has
been moved to the initActionListeners() method and button parameter
settings has been moved to the setButtonParameters() method, both of
which are called from initControls().
Added unit selection buttons, as well as their appropriate settings in
the initActionListeners() and setButtonParameters() methods.  (not
functional yet).
Added new labels to display unit power/strength.
added setLabelParameters() method to set the parameters of the newly
introduced labels.

CONQUESTMAP.JAVA CHANGES
Added soldier units.
Added a check to ensure that ConquestMap.java is not run directly (i.e.
from the main method), as it needs to be called from Conquest.java to
run properly.
Replaced the moveRight() method with move() method. Now supports moving
in all 4 directions.
Made some methods private, as they are not needed globally.

FILEACCESSOR.JAVA CHANGES
Removed a JOptionPane.showMessageDialog(etc.) line in the msgBox()
method, as it caused the creation of duplicate dialog boxes, which was
not intended.