Alpha 0.6.1
CHANGES TO CONQUEST.JAVA
Methods have been reorganised to make it more readable.

CHANGES TO CONQUESTMAP.JAVA
Units now move using a coordinate system that correspond to the relevant
location on the screen, laying down the basic structure for a grid
management system, borders and control points in a later update.
Renamed initGridOccupied() method to resetGridOccupied() to better
reflect its purpose.
Added setGridOccupied() method. Changes the relevant coordinates for the
specified unit defined in the parameters and set the relevant
coordinate's occupation status to true. Only functional for Player 1
unit 1 for now.