Release Candidate 2
Updated data dictionary
Made numbers on soldier units white with black stroke for better
visibility
Set up foundation for HowToPlay class. How to play instructions will be
coming in the next release candidate.
Added "all units killed" labels to ConquestScoresheet which appear if
all of a player's units are killed.
Fixed bug where move buttons were enabled if a unit was upgraded, even
if no unit was selected.
It is now possible to win by killing all of the enemy's units.
Fixed capture points C and D not corresponding to their appropriate
locations on the map.
Added obstructions to the map.