Release Candidate 7
This is a major update to address an issue where a requirement that the
software solution "Writes to file and saves game score" was not met. As
a gesture of goodwill, it has been decided to no only address this
issue, but also include a function to save the game's progress so that
players may resume playing later.

CHANGES TO CONQUEST AS A WHOLE
Conquest now uses the FileAccessor API (written by Weilon)

CHANGES TO CONQUEST.JAVA
Implemented score system. Score is given to the winner of the game and
is calculated by: score = number of own units alive * number of enemy
units dead * tickets the winning player has left
Introduced new global string variables: stringSaveFile and
stringHighScoreFile. These define the file names of save files that will
be used to store data in the long term.
Added additional comments for more readability
Pressing btnEndTurn (End Turn Button) without having used all the moves
and rolled the dice will now prompt the user to ensure he/she knows what
they are doing.
Added additional parameter to processWin method to accommodate the score
system.
Added calculateScore() method. Calculates the score according to the
algorithm as shown above.
Exiting the game now prompts the user whether they would like to save
the game or not.
Added saveGame() method. Writes all necessary variables to file for long
term storage.
Added checkGameSave() method. Checks for an existing save file, then
asks the user if they would like to restore that game if it is
incomplete. If user says yes, it will restore the variables to its
previous state when the game was saved, then repositions the units.
Added reloadGame() method. Forces GUI elements to refresh to reflect the
updated variables when restoring from a save.

CHANGES TO CONQUESTMAP.JAVA
Attempting to close the game through ConquestMap now tells the user to
close through the control window instead.

CHANGES TO CONQUESTSTART.JAVA
Window now displays the current high score.
Added getHighScores() method. Used to obtain the high score and the
player associated with it, then displays it on the screen.

CHANGES TO FILEACCESSOR.JAVA
The FileAccessor API has been customised to better suit the software
solution.
read() method now throws Exceptions
regenFile() method now detects if it is being used to regenerate the
high score or save file, and creates the appropriate properties through
the use of the new methods writeHighScore() and
writeConquestProperties();
Added deleteFile() method. It is used to delete files (obviously).